+++ src/TestSOM.java
@@ -10,14 +10,14 @@ import com.wiseaux.setCreator.SetCreatorEnhanced;
 public class TestSOM {
     
     public static void main(String[] args){
-//        SetCreatorEnhanced som = new SetCreatorEnhanced();
-//        som.runForPreCreatedSets();
-//        som.trainPreCreatedSets();
-        Intersector thing = new Intersector();
-        thing.read(10);
-        thing.run();
-        //thing.resultsToFile("result");
-        thing.totalResultsFile();
+        SetCreatorEnhanced som = new SetCreatorEnhanced();
+        som.runForSpecialSet();
+        som.train(10);
+//        Intersector thing = new Intersector();
+//        thing.read(10);
+//        thing.run();
+//        //thing.resultsToFile("result");
+//        thing.totalResultsFile();
     }
         
 }