+++ src/com/wiseaux/setCreator/RandRecordSet.java
@@ -23,7 +23,7 @@ public class RandRecordSet {
      * Holds all records from the input file.
      */
     private Record[] allRecords;
-
+    
     /**
      * Holds all records with the first of two decision values.
      */
@@ -178,6 +178,22 @@ public class RandRecordSet {
 
         return groups;
     }
+    
+    /**
+     * Returns a queue of 426 unique groups that contain a singleRecord aRecord.
+     */
+    public Queue[] createSpecialTestSet(Record[] records, int decVal_0, int decVal_1) {
+        splitRecords(records, decVal_0, decVal_1);
+        Queue[] groups = new Queue[records.length];
+        int i = 0;
+        for(Record aRecord : records) {
+            Queue<Record> singleRecord = new LinkedList<>();
+            singleRecord.add(aRecord);
+            groups[i++] = singleRecord;
+        }
+        
+        return groups;
+    }
 
     /**
      * Grab random records from the queue equal to 10% of the length of the
@@ -191,7 +207,7 @@ public class RandRecordSet {
 
         return tenSet;
     }
-
+    
     /**
      * Split allRecords into records with decision 0 and decision 1.
      */
@@ -211,23 +227,23 @@ public class RandRecordSet {
             }
         }
     }
-
+    
     /**
      * Takes values that aren't in the current testSet and returns an equal
      * amount of records from decision_0 and decision_1.
      */
-    public Queue[] createTrainingSet(Queue[] groups) {
-        Queue[] trainingGroups = new Queue[20];
+    public Queue[] createTrainingSet(Queue[] groups, double percent) {
+        Queue[] trainingGroups = new Queue[groups.length];
 
-        int size = getSizeDifference(decision_0, decision_1);
+        int size = getSizeDifference(decision_0, decision_1, percent);
         int i = 0;
         for (Queue<Record> queue : groups) {
-            if(i >= 10) {
-                Queue<Record> newTrainingSet = getLeftovers(decision_1, queue, size, i/10);
+            if(i >= (groups.length/2)) {
+                Queue<Record> newTrainingSet = getLeftovers(decision_1, queue, size);
                 trainingGroups[i] = newTrainingSet;
                 
             } else {
-                Queue<Record> newTrainingSet = getLeftovers(decision_0, queue, size, i/10);
+                Queue<Record> newTrainingSet = getLeftovers(decision_0, queue, size);
                 trainingGroups[i] = newTrainingSet;
             }
             i++;
@@ -239,7 +255,7 @@ public class RandRecordSet {
     /**
      * Gets the leftover records from the given group.
      */
-    private Queue<Record> getLeftovers(Queue<Record> decisionSet, Queue<Record> testSet, int length, int dec) {
+    private Queue<Record> getLeftovers(Queue<Record> decisionSet, Queue<Record> testSet, int length) {
         Queue<Record> leftovers = new LinkedList<>();
         
         for(Record decRecord: decisionSet) {
@@ -267,14 +283,14 @@ public class RandRecordSet {
      * Gets the sizes of the two queues and if one is larger than the other,
      * take the smaller of the two.
      */
-    private int getSizeDifference(Queue<Record> dec_0, Queue<Record> dec_1) {
-        Double value_1 = dec_0.size() - (dec_0.size() * 0.1);
-        Double value_2 = dec_1.size() - (dec_1.size() * 0.1);
+    private int getSizeDifference(Queue<Record> dec_0, Queue<Record> dec_1, double percent) {
+        Double value_1 = dec_0.size() - (dec_0.size() * percent);
+        Double value_2 = dec_1.size() - (dec_1.size() * percent);
         if (value_1 > value_2) {
             return value_2.intValue();
-        } else {
-            return value_1.intValue();
         }
+        return value_1.intValue();
+        
     }
     
     public void printQue(Queue<Record> que, String name){